+++ ResultFunctional/FunctionalExtensions/Async/CurryTaskAsyncExtensions.cs
@@ -1,25 +1,28 @@
 ﻿using System;
+using System.Threading.Tasks;
+using ResultFunctional.FunctionalExtensions.Sync;
 
-namespace ResultFunctional.FunctionalExtensions.Sync
+namespace ResultFunctional.FunctionalExtensions.Async
 {
     /// <summary>
-    /// Higher order functions
+    /// Task higher order functions
     /// </summary>
-    public static class CurryAsyncExtensions
+    public static class CurryTaskAsyncExtensions
     {
         /// <summary>
-        /// Get no arguments higher order function
+        /// Get no arguments task higher order function
         /// </summary>
         /// <typeparam name="TIn1">Argument type</typeparam>
         /// <typeparam name="TOut">Function type</typeparam>
         /// <param name="this">One argument higher order function</param>
         /// <param name="arg1">Argument</param>
         /// <returns>No arguments higher order function</returns>
-        public static Func<TOut> Curry<TIn1, TOut>(this Func<TIn1, TOut> @this, TIn1 arg1) =>
-            () => @this(arg1);
+        public static async Task<Func<TOut>> CurryTaskAsync<TIn1, TOut>(this Task<Func<TIn1, TOut>> @this, TIn1 arg1) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.Curry(arg1));
 
         /// <summary>
-        /// Get one argument higher order function
+        /// Get one argument task higher order function
         /// </summary>
         /// <typeparam name="TIn1">Argument type</typeparam>
         /// <typeparam name="TIn2">Argument type</typeparam>
@@ -27,11 +30,13 @@ namespace ResultFunctional.FunctionalExtensions.Sync
         /// <param name="this">Two arguments higher order function</param>
         /// <param name="arg1">Argument</param>
         /// <returns>One argument higher order function</returns>
-        public static Func<TIn2, TOut> Curry<TIn1, TIn2, TOut>(this Func<TIn1, TIn2, TOut> @this, TIn1 arg1) =>
-            arg2 => @this(arg1, arg2);
+        public static async Task<Func<TIn2, TOut>> CurryTaskAsync<TIn1, TIn2, TOut>(this Task<Func<TIn1, TIn2, TOut>> @this,
+                                                                                TIn1 arg1) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.Curry(arg1));
 
         /// <summary>
-        /// Get two arguments higher order function
+        /// Get two arguments task higher order function
         /// </summary>
         /// <typeparam name="TIn1">Argument type</typeparam>
         /// <typeparam name="TIn2">Argument type</typeparam>
@@ -40,11 +45,13 @@ namespace ResultFunctional.FunctionalExtensions.Sync
         /// <param name="this">Three arguments higher order function</param>
         /// <param name="arg1">Argument</param>
         /// <returns>Two argument higher order function</returns>
-        public static Func<TIn2, TIn3, TOut> Curry<TIn1, TIn2, TIn3, TOut>(this Func<TIn1, TIn2, TIn3, TOut> @this, TIn1 arg1) =>
-            (arg2, arg3) => @this(arg1, arg2, arg3);
+        public static async Task<Func<TIn2, TIn3, TOut>> CurryTaskAsync<TIn1, TIn2, TIn3, TOut>(this Task<Func<TIn1, TIn2, TIn3, TOut>> @this,
+                                                                                TIn1 arg1) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.Curry(arg1));
 
         /// <summary>
-        /// Get three arguments higher order function
+        /// Get three arguments task higher order function
         /// </summary>
         /// <typeparam name="TIn1">Argument type</typeparam>
         /// <typeparam name="TIn2">Argument type</typeparam>
@@ -54,12 +61,13 @@ namespace ResultFunctional.FunctionalExtensions.Sync
         /// <param name="this">Four arguments higher order function</param>
         /// <param name="arg1">Argument</param>
         /// <returns>Three argument higher order function</returns>
-        public static Func<TIn2, TIn3, TIn4, TOut> Curry<TIn1, TIn2, TIn3, TIn4, TOut>(this Func<TIn1, TIn2, TIn3, TIn4, TOut> @this,
-                                                                                       TIn1 arg1) =>
-            (arg2, arg3, arg4) => @this(arg1, arg2, arg3, arg4);
+        public static async Task<Func<TIn2, TIn3, TIn4, TOut>> CurryTaskAsync<TIn1, TIn2, TIn3, TIn4, TOut>(this Task<Func<TIn1, TIn2, TIn3, TIn4, TOut>> @this,
+                                                                                            TIn1 arg1) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.Curry(arg1));
 
         /// <summary>
-        /// Get four arguments higher order function
+        /// Get four arguments task higher order function
         /// </summary>
         /// <typeparam name="TIn1">Argument type</typeparam>
         /// <typeparam name="TIn2">Argument type</typeparam>
@@ -70,8 +78,9 @@ namespace ResultFunctional.FunctionalExtensions.Sync
         /// <param name="this">Five arguments higher order function</param>
         /// <param name="arg1">Argument</param>
         /// <returns>Four argument higher order function</returns>
-        public static Func<TIn2, TIn3, TIn4, TIn5, TOut> Curry<TIn1, TIn2, TIn3, TIn4, TIn5, TOut>(this Func<TIn1, TIn2, TIn3, TIn4, TIn5, TOut> @this,
-                                                                                                   TIn1 arg1) =>
-            (arg2, arg3, arg4, arg5) => @this(arg1, arg2, arg3, arg4, arg5);
+        public static async Task<Func<TIn2, TIn3, TIn4, TIn5, TOut>> CurryTaskAsync<TIn1, TIn2, TIn3, TIn4, TIn5, TOut>(this Task<Func<TIn1, TIn2, TIn3, TIn4, TIn5, TOut>> @this,
+                                                                                                        TIn1 arg1) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.Curry(arg1));
     }
 }
