+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindTryAsyncExtensions.cs
@@ -7,13 +7,17 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Методы расширения для результирующего ответа со связыванием со значением и обработкой исключений асинхронно
+    /// Extension methods for result value async monad function with exception handling
     /// </summary>
     public static class ResultValueBindTryAsyncExtensions
     {
         /// <summary>
-        /// Обработать асинхронную функцию, вернуть результирующий ответ со связыванием со значением или ошибку исключения
+        /// Execute result value async monad function with exception handling
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="func">Result value function</param>
+        /// <param name="exceptionFunc">Exception function</param>
+        /// <returns>Result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindTryAsync<TValue>(Func<Task<IResultValue<TValue>>> func,
                                                                                        Func<Exception, IErrorResult> exceptionFunc)
         {
@@ -28,8 +32,12 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
         }
 
         /// <summary>
-        /// Обработать асинхронную функцию, вернуть результирующий ответ со связыванием со значением или ошибку исключения
+        /// Execute result collection async monad function with exception handling
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="func">Result value function</param>
+        /// <param name="error">Error</param>
+        /// <returns>Result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindTryAsync<TValue>(Func<Task<IResultValue<TValue>>> func,
                                                                                        IErrorResult error) =>
              await ResultValueBindTryAsync(func, error.AppendException);
