+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindTryWhereTaskAsyncExtensions.cs
@@ -8,13 +8,19 @@ using static ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultVa
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Методы расширения для результирующего ответа со связыванием со значением и обработкой исключений для задач-объектов
+    /// Extension methods for task result value monad function with conditions and exception handling
     /// </summary>
     public static class ResultValueBindTryWhereTaskAsyncExtensions
     {
         /// <summary>
-        /// Результирующий ответ cj связыванием со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute task result value function in no error case; else catch exception
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Monad result value function</param>
+        /// <param name="exceptionFunc">Exception function</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueBindTryOkTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                        Func<TValueIn, IResultValue<TValueOut>> func,
                                                                                                        Func<Exception, IErrorResult> exceptionFunc) =>
@@ -22,8 +28,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             ResultValueBindOkTaskAsync(value => ResultValueBindTry(() => func.Invoke(value), exceptionFunc));
 
         /// <summary>
-        /// Результирующий ответ cj связыванием со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute task result value function in no error case; else catch exception
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Monad result value function</param>
+        /// <param name="error">Error</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueBindTryOkTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                        Func<TValueIn, IResultValue<TValueOut>> func,
                                                                                                        IErrorResult error) =>
