+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindWhereAsyncExtensions.cs
@@ -10,13 +10,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка условий для результирующего асинхронного связывающего ответа со значением
+    /// Extension methods for result value async monad function with conditions
     /// </summary>
     public static class ResultCollectionBindWhereAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в асинхронном результирующем ответе со значением
-        /// </summary>      
+        /// Execute monad result value async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>     
         public static async Task<IResultValue<TValueOut>> ResultValueBindContinueAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                             Func<TValueIn, bool> predicate,
                                                                                                             Func<TValueIn, Task<IResultValue<TValueOut>>> okFunc,
@@ -28,8 +35,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
                 : new ResultValue<TValueOut>(@this.Errors);
 
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в асинхронном результирующем ответе со значением
-        /// </summary>      
+        /// Execute monad result value async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>     
         public static async Task<IResultValue<TValueOut>> ResultValueBindWhereAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                          Func<TValueIn, bool> predicate,
                                                                                                          Func<TValueIn, Task<IResultValue<TValueOut>>> okFunc,
@@ -41,8 +55,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
                 : new ResultValue<TValueOut>(@this.Errors);
 
         /// <summary>
-        /// Выполнение положительного или негативного условия в асинхронном результирующем ответе со значением
-        /// </summary>      
+        /// Execute monad result value async function depending on result value errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if result value hasn't errors</param>
+        /// <param name="badFunc">Function if result value has errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueBindOkBadAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                          Func<TValueIn, Task<IResultValue<TValueOut>>> okFunc,
                                                                                                          Func<IReadOnlyCollection<IErrorResult>, Task<IResultValue<TValueOut>>> badFunc) =>
@@ -50,9 +70,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
                 ? await okFunc.Invoke(@this.Value)
                 : await badFunc.Invoke(@this.Errors);
 
+
         /// <summary>
-        /// Выполнение положительного условия результирующего асинхронного ответа со связыванием или возвращение предыдущей ошибки в результирующем ответе
-        /// </summary>   
+        /// Execute monad result value async function if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueBindOkAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                       Func<TValueIn, Task<IResultValue<TValueOut>>> okFunc) =>
             @this.OkStatus
@@ -60,8 +86,12 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
                 : new ResultValue<TValueOut>(@this.Errors);
 
         /// <summary>
-        /// Выполнение негативного условия результирующего асинхронного ответа или возвращение положительного в результирующем ответе
-        /// </summary>   
+        /// Execute monad result value async function if incoming result value has errors
+        /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="badFunc">Function if incoming result value has errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindBadAsync<TValue>(this IResultValue<TValue> @this,
                                                                                        Func<IReadOnlyCollection<IErrorResult>, Task<IResultValue<TValue>>> badFunc) =>
             @this.OkStatus
@@ -69,8 +99,12 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
                 : await badFunc.Invoke(@this.Errors);
 
         /// <summary>
-        /// Добавить асинхронно ошибки результирующего ответа или вернуть результат с ошибками для ответа со значением
+        /// Adding errors async to result value if ones hasn't errors
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Error function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindErrorsOkAsync<TValue>(this IResultValue<TValue> @this,
                                                                                             Func<TValue, Task<IResultError>> okFunc) =>
             await @this.
