+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindWhereTaskAsyncExtensions.cs
@@ -9,13 +9,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка условий для результирующего связывающего ответа со значением для задачи-объекта
+    /// Extension methods for task result value monad function with conditions
     /// </summary>
     public static class ResultValueBindWhereTaskAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе со значением для задачи-объекта
-        /// </summary>      
+        /// Execute task monad result value function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>        
         public static async Task<IResultValue<TValueOut>> ResultValueBindContinueTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                                 Func<TValueIn, bool> predicate,
                                                                                                                 Func<TValueIn, IResultValue<TValueOut>> okFunc,
@@ -24,8 +31,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapTaskAsync(awaitedThis => awaitedThis.ResultValueBindContinue(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе со значением для задачи-объекта
-        /// </summary>      
+        /// Execute task monad result value function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>          
         public static async Task<IResultValue<TValueOut>> ResultValueBindWhereTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                              Func<TValueIn, bool> predicate,
                                                                                                              Func<TValueIn, IResultValue<TValueOut>> okFunc,
@@ -34,8 +48,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapTaskAsync(awaitedThis => awaitedThis.ResultValueBindWhere(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного или негативного условия в результирующем ответе со значением для задачи-объекта
-        /// </summary>      
+        /// Execute task monad result value function depending on result value errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if result value hasn't errors</param>
+        /// <param name="badFunc">Function if result value has errors</param>
+        /// <returns>Outgoing result value</returns>     
         public static async Task<IResultValue<TValueOut>> ResultValueBindOkBadTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                              Func<TValueIn, IResultValue<TValueOut>> okFunc,
                                                                                                              Func<IReadOnlyCollection<IErrorResult>, IResultValue<TValueOut>> badFunc) =>
@@ -43,24 +63,37 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapTaskAsync(awaitedThis => awaitedThis.ResultValueBindOkBad(okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного условия результирующего ответа или возвращение предыдущей ошибки в результирующем ответе для задачи-объекта
-        /// </summary>   
+        /// Execute task monad result value function if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns> 
         public static async Task<IResultValue<TValueOut>> ResultValueBindOkTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                           Func<TValueIn, IResultValue<TValueOut>> okFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultValueBindOk(okFunc));
 
         /// <summary>
-        /// Выполнение негативного условия результирующего ответа или возвращение положительного в результирующем ответе для задачи-объекта
-        /// </summary>   
+        /// Execute task monad result value function if incoming result value has errors
+        /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="badFunc">Function if incoming result value has errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindBadTaskAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                            Func<IReadOnlyCollection<IErrorResult>, IResultValue<TValue>> badFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultValueBindBad(badFunc));
 
         /// <summary>
-        /// Добавить ошибки результирующего ответа или вернуть результат с ошибками для ответа со значением для задачи-объекта
+        /// Adding errors to task result value if ones hasn't errors
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Error function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBindErrorsOkTaskAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                             Func<TValue, IResultError> okFunc) =>
             await @this.
