+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindWhereToCollectionAsyncExtensions.cs
@@ -7,13 +7,18 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка условий для асинхронного результирующего ответа с значением с возвращением к коллекции
+    /// Extension methods for result value async monad function converting to result collection
     /// </summary>
     public static class ResultValueBindWhereToCollectionAsyncExtensions
     {
         /// <summary>
-        /// Выполнение асинхронного положительного условия или возвращение предыдущей ошибки в результирующем ответе значения с возвращением к коллекции
-        /// </summary>   
+        /// Execute monad result value async function converting to result collection if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultValueBindOkToCollectionAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                              Func<TValueIn, Task<IResultCollection<TValueOut>>> okFunc) =>
             await @this.
