+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueBindWhereToCollectionTaskAsyncExtensions.cs
@@ -6,13 +6,18 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка условий для результирующего ответа задачи-объекта с значением с возвращением к коллекции
+    /// Extension methods for task result value monad function converting to result collection
     /// </summary>
     public static class ResultValueBindWhereToCollectionTaskAsyncExtensions
     {
         /// <summary>
-        /// Выполнение асинхронного положительного условия или возвращение предыдущей ошибки в результирующем ответе значения с возвращением к коллекции
-        /// </summary>   
+        /// Execute monad task result value function converting to result collection if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if incoming result value hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultValueBindOkToCollectionTaskAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                              Func<TValueIn, IResultCollection<TValueOut>> okFunc) =>
             await @this.
