+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueToValueBindAsyncExtensions.cs
@@ -8,13 +8,19 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Преобразование результирующего ответа в значение асинхронно для задачи-объекта
+    /// Extension methods for task value async function converting to value
     /// </summary>
     public static class ResultValueToValueBindAsyncExtensions
     {
         /// <summary>
-        /// Выполнение положительного или негативного условия в результирующем ответе и преобразование в значение
-        /// </summary>      
+        /// Execute task value async function converting to another value if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if incoming result value hasn't errors</param>
+        /// <param name="badFunc">Function if incoming result value has errors</param>
+        /// <returns>Outgoing value</returns>
         public static async Task<TValueOut> ResultValueToValueOkBadBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                     Func<TValueIn, Task<TValueOut>> okFunc,
                                                                                     Func<IReadOnlyCollection<IErrorResult>, Task<TValueOut>> badFunc) =>
