+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueTryWhereAsyncExtensions.cs
@@ -8,13 +8,19 @@ using static ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultV
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Методы расширения для результирующего ответа со значением и обработкой исключений асинхронно
+    /// Async exception handling result value with conditions extension methods
     /// </summary>
     public static class ResultValueTryWhereAsyncExtensions
     {
         /// <summary>
-        /// Результирующий ответ со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute async function and handle exception with result value concat
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Value function</param>
+        /// <param name="exceptionFunc">Function converting exception to error</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueTryOkAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                      Func<TValueIn, Task<TValueOut>> func,
                                                                                                      Func<Exception, IErrorResult> exceptionFunc) =>
@@ -22,8 +28,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             ResultValueBindOkAsync(value => ResultValueTryAsync(() => func.Invoke(value), exceptionFunc));
 
         /// <summary>
-        /// Результирующий ответ со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute async function and handle exception with result value concat
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Value function</param>
+        /// <param name="error">Error</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueTryOkAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                      Func<TValueIn, Task<TValueOut>> func,
                                                                                                      IErrorResult error) =>
