+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueTryWhereBindAsyncExtensions.cs
@@ -8,13 +8,19 @@ using static ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultV
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Методы расширения для результирующего ответа со значением и обработкой исключений
+    /// Async exception handling task result value with conditions extension methods
     /// </summary>
     public static class ResultValueTryWhereBindAsyncExtensions
     {
         /// <summary>
-        /// Связать результирующий ответ со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute async function and handle exception with task result value concat
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Value function</param>
+        /// <param name="exceptionFunc">Function converting exception to error</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueTryOkBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                          Func<TValueIn, Task<TValueOut>> func,
                                                                                                          Func<Exception, IErrorResult> exceptionFunc) =>
@@ -22,8 +28,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             ResultValueBindOkBindAsync(value => ResultValueTryAsync(() => func.Invoke(value), exceptionFunc));
 
         /// <summary>
-        /// Связать результирующий ответ со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute async function and handle exception with task result value concat
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="func">Value function</param>
+        /// <param name="error">Error</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ResultValueTryOkBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                          Func<TValueIn, Task<TValueOut>> func,
                                                                                                          IErrorResult error) =>
