+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueVoidBindAsyncExtensions.cs
@@ -7,29 +7,42 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Асинхронное действие над внутренним типом результирующего ответа со значением задачей-объектом
+    /// Task result value async action extension methods
     /// </summary>
     public static class ResultCollectionVoidBindAsyncExtensions
     {
         /// <summary>
-        /// Выполнить действие при положительном значении, вернуть результирующий ответ задачи-объекта
-        /// </summary>      
+        /// Execute async action if task result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValue">Incoming type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="action">Action</param>
+        /// <returns>Unchanged result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueVoidOkBindAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                      Func<TValue, Task> action) =>
             await @this.
             MapBindAsync(awaitedThis => awaitedThis.ResultValueVoidOkAsync(action));
 
         /// <summary>
-        /// Выполнить действие при отрицательном значении, вернуть результирующий ответ задачи-объекта
-        /// </summary>      
+        /// Execute async action if task result value has errors
+        /// </summary>
+        /// <typeparam name="TValue">Incoming type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="action">Action</param>
+        /// <returns>Unchanged result value</returns> 
         public static async Task<IResultValue<TValue>> ResultValueVoidBadBindAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                       Func<IReadOnlyCollection<IErrorResult>, Task> action) =>
             await @this.
             MapBindAsync(awaitedThis => awaitedThis.ResultValueVoidBadAsync(action));
 
         /// <summary>
-        /// Выполнить действие, вернуть результирующий ответ задачи-объекта
-        /// </summary>      
+        /// Execute async action depending on task result value errors
+        /// </summary>
+        /// <typeparam name="TValue">Incoming type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="actionOk">Action if result value hasn't errors</param>
+        /// <param name="actionBad">Action if result value has errors</param>
+        /// <returns>Unchanged result value</returns>  
         public static async Task<IResultValue<TValue>> ResultValueVoidOkBadBindAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                              Func<TValue, Task> actionOk,
                                                                                              Func<IReadOnlyCollection<IErrorResult>, Task> actionBad) =>
@@ -37,8 +50,13 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(awaitedThis => awaitedThis.ResultValueVoidOkBadAsync(actionOk, actionBad));
 
         /// <summary>
-        /// Выполнить действие при положительном значении и выполнении условия вернуть результирующий ответ задачи-объекта
-        /// </summary>    
+        /// Execute async action depending on task result value errors and predicate
+        /// </summary>
+        /// <typeparam name="TValue">Incoming type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="action">Function if predicate <see langword="true"/></param>
+        /// <returns>Unchanged result value</returns>   
         public static async Task<IResultValue<TValue>> ResultValueVoidOkWhereBindAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                           Func<TValue, bool> predicate,
                                                                                           Func<TValue, Task> action) =>
