+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueWhereBindAsyncExtensions.cs
@@ -8,13 +8,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка асинхронных условий для результирующего ответа со значением задачей-объектом
+    /// Extension methods for task result value async functor function with conditions
     /// </summary>
     public static class ResultValueWhereTaskBindAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в асинхронном результирующем ответе задачи-объекта
-        /// </summary>      
+        /// Execute task result value async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns> 
         public static async Task<IResultValue<TValueOut>> ResultValueContinueBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                             Func<TValueIn, bool> predicate,
                                                                                                             Func<TValueIn, Task<TValueOut>> okFunc,
@@ -23,8 +30,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(awaitedThis => awaitedThis.ResultValueContinueAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение условия в положительном или негативном варианте в результирующем ответе задачи-объекта
-        /// </summary>      
+        /// Execute task result value async function base on predicate condition returning value in any case
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>   
         public static async Task<IResultValue<TValueOut>> ResultValueWhereBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                          Func<TValueIn, bool> predicate,
                                                                                                          Func<TValueIn, Task<TValueOut>> okFunc,
@@ -33,8 +47,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(awaitedThis => awaitedThis.ResultValueWhereAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного или негативного условия в асинхронном результирующем ответе задачи-объекта
-        /// </summary>      
+        /// Execute task result value async function depending on result value errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if result value hasn't errors</param>
+        /// <param name="badFunc">Function if result value has errors</param>
+        /// <returns>Outgoing result value</returns> 
         public static async Task<IResultValue<TValueOut>> ResultValueOkBadBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                          Func<TValueIn, Task<TValueOut>> okFunc,
                                                                                                          Func<IReadOnlyCollection<IErrorResult>, Task<TValueOut>> badFunc) =>
@@ -42,16 +62,25 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(awaitedThis => awaitedThis.ResultValueOkBadAsync(okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного условия или возвращение предыдущей ошибки в асинхронном результирующем ответе задачи-объекта
-        /// </summary>   
+        /// Execute task result value async function if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns>  
         public static async Task<IResultValue<TValueOut>> ResultValueOkBindAsync<TValueIn, TValueOut>(this Task<IResultValue<TValueIn>> @this,
                                                                                                       Func<TValueIn, Task<TValueOut>> okFunc) =>
             await @this.
             MapBindAsync(awaitedThis => awaitedThis.ResultValueOkAsync(okFunc));
 
         /// <summary>
-        /// Выполнение негативного условия или возвращение положительного условия в асинхронном результирующем ответе задачи-объекта
-        /// </summary>   
+        /// Execute task result value async function if incoming result value has errors
+        /// </summary>
+        /// <typeparam name="TValue">Incoming type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="badFunc">Function if result value has errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ResultValueBadBindAsync<TValue>(this Task<IResultValue<TValue>> @this,
                                                                                        Func<IReadOnlyCollection<IErrorResult>, Task<TValue>> badFunc) =>
             await @this.
