+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueWhereToCollectionAsyncExtensions.cs
@@ -8,13 +8,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Обработка условий для асинхронного результирующего ответа с значением с возвращением к коллекции
+    /// Extension methods for result value async functions converting to result value
     /// </summary>
     public static class ResultValueWhereToCollectionAsyncExtensions
     {
         /// <summary>
-        /// Выполнение асинхронного условия или возвращение предыдущей ошибки в результирующем ответе значения с возвращением к коллекции
-        /// </summary>      
+        /// Execute result value async function converting to result value base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>  
         public static async Task<IResultCollection<TValueOut>> ResultValueContinueToCollectionAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                                                         Func<TValueIn, bool> predicate,
                                                                                                                         Func<TValueIn, Task<IEnumerable<TValueOut>>> okFunc,
@@ -24,8 +31,14 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             ToResultCollectionTaskAsync();
 
         /// <summary>
-        /// Выполнение асинхронного положительного или негативного условия в результирующем ответе значения с возвращением к коллекции
-        /// </summary>      
+        /// Execute result value async function converting to result value depending on result value errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>     
         public static async Task<IResultCollection<TValueOut>> ResultValueOkBadToCollectionAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                     Func<TValueIn, Task<IEnumerable<TValueOut>>> okFunc,
                                                                                     Func<IReadOnlyCollection<IErrorResult>, Task<IEnumerable<TValueOut>>> badFunc) =>
@@ -34,8 +47,13 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             ToResultCollectionTaskAsync();
 
         /// <summary>
-        /// Выполнение асинхронного положительного условия или возвращение предыдущей ошибки в результирующем ответе значения с возвращением к коллекции
-        /// </summary>   
+        /// Execute result value async function converting to result value if incoming result value hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result value</param>
+        /// <param name="okFunc">Function if result value hasn't errors</param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultCollection<TValueOut>> ResultValueOkToCollectionAsync<TValueIn, TValueOut>(this IResultValue<TValueIn> @this,
                                                                                              Func<TValueIn, Task<IEnumerable<TValueOut>>> okFunc) =>
             await  @this.
