+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ToResultValueWhereBindAsyncExtensions.cs
@@ -6,13 +6,18 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues
 {
     /// <summary>
-    /// Преобразование значения в результирующий ответ с условием асинхронно для задачи-объекта
+    /// Task result value async extension methods with condition
     /// </summary>
     public static class ToResultValueWhereBindAsyncExtensions
     {
         /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием
+        /// Async converting task value to result value base on predicate
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="badFunc">Error function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ToResultValueWhereBindAsync<TValue>(this Task<TValue> @this,
                                                                                            Func<TValue, bool> predicate,
                                                                                            Func<TValue, Task<IErrorResult>> badFunc)
@@ -21,8 +26,13 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
           MapBindAsync(thisAwaited => thisAwaited.ToResultValueWhereAsync(predicate, badFunc));
 
         /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием и проверкой на нуль
+        /// Async converting task value to result value base on predicate
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="badFunc">Error function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ToResultValueWhereNullBindAsync<TValue>(this Task<TValue?> @this,
                                                                                       Func<TValue, bool> predicate,
                                                                                       Func<TValue?, Task<IErrorResult>> badFunc)
@@ -31,8 +41,13 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(thisAwaited => thisAwaited.ToResultValueWhereNullAsync(predicate, badFunc));
 
         /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием и проверкой на нуль
+        /// Async converting task value to result value base on predicate
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="badFunc">Error function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValue>> ToResultValueWhereNullBindAsync<TValue>(this Task<TValue?> @this,
                                                                                       Func<TValue, bool> predicate,
                                                                                       Func<TValue?, Task<IErrorResult>> badFunc)
@@ -41,8 +56,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(thisAwaited => thisAwaited.ToResultValueWhereNullAsync(predicate, badFunc));
 
         /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием и проверкой на нуль
+        /// Async converting task value to result value base on predicate base with functor function
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Functor function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Error function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ToResultValueWhereNullOkBadBindAsync<TValueIn, TValueOut>(this Task<TValueIn?> @this,
                                                                                       Func<TValueIn, bool> predicate,
                                                                                       Func<TValueIn, Task<TValueOut>> okFunc,
@@ -53,8 +75,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
             MapBindAsync(thisAwaited => thisAwaited.ToResultValueWhereNullOkBadAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием и проверкой на нуль
+        /// Async converting task value to result value base on predicate base with functor function
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming value</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Functor function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Error function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result value</returns>
         public static async Task<IResultValue<TValueOut>> ToResultValueWhereNullOkBadBindAsync<TValueIn, TValueOut>(this Task<TValueIn?> @this,
                                                                                       Func<TValueIn, bool> predicate,
                                                                                       Func<TValueIn, Task<TValueOut>> okFunc,
