+++ ResultFunctional/FunctionalExtensions/Sync/ResultExtension/ResultErrors/ToResultErrorExtensions.cs
@@ -26,29 +26,5 @@ namespace ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultError
         /// <returns>Result error</returns>
         public static IResultError ToResultError(this IEnumerable<IErrorResult> @this) =>
             new ResultError(@this);
-
-        /// <summary>
-        /// Converting result error to result value
-        /// </summary>
-        /// <typeparam name="TValue">Result type</typeparam>
-        /// <param name="this">Incoming result error</param>
-        /// <param name="value">Value</param>
-        /// <returns>Outgoing result value</returns>
-        public static IResultValue<TValue> ToResultValue<TValue>(this IResultError @this, TValue value) =>
-            @this.OkStatus
-                ? new ResultValue<TValue>(value)
-                : new ResultValue<TValue>(@this.Errors);
-
-        /// <summary>
-        /// Merge result error with result value
-        /// </summary>
-        /// <typeparam name="TValue">Result type</typeparam>
-        /// <param name="this">Incoming result error</param>
-        /// <param name="resultValue">Result value</param>
-        /// <returns>Outgoing result value</returns>
-        public static IResultValue<TValue> ToResultBindValue<TValue>(this IResultError @this, IResultValue<TValue> resultValue) =>
-            @this.OkStatus
-                ? resultValue
-                : new ResultValue<TValue>(@this.Errors);
     }
 }
