--- ResultFunctionalXUnit/FunctionalExtensions/Async/ResultExtension/ResultErrorTest/ToResultErrorBindAsyncExtensionsTest.cs
@@ -1,88 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultErrors;
-using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultValues;
-using ResultFunctional.Models.Implementations.ResultFactory;
-using ResultFunctional.Models.Implementations.Results;
-using ResultFunctional.Models.Interfaces.Errors.Base;
-using ResultFunctional.Models.Interfaces.Results;
-using ResultFunctionalXUnit.Data;
-using Xunit;
-
-namespace ResultFunctionalXUnit.FunctionalExtensions.Async.ResultExtension.ResultErrorTest
-{
-    /// <summary>
-    /// Методы расширения для результирующего ответа для задачи-объекта. Тесты
-    /// </summary>
-    public class ToResultErrorBindAsyncExtensionsTest
-    {
-        /// <summary>
-        /// Вернуть результирующий ответ со значением без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueBindAsync_OkStatus()
-        {
-            var resultNoError = ResultErrorFactory.CreateTaskResultError();
-            const string value = "OkStatus";
-            var resultValue = ResultValueFactory.CreateTaskResultValue(value);
-
-            var resultValueAfter = await resultNoError.ToResultBindValueBindAsync(resultValue);
-
-            Assert.True(resultValueAfter.OkStatus);
-            Assert.Equal(value, resultValueAfter.Value);
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueBindAsync_HasErrors()
-        {
-            var error = ErrorData.CreateErrorTest();
-            var resultHasError = ResultErrorFactory.CreateTaskResultError(error);
-            const string value = "BadStatus";
-            var resultValue = ResultValueFactory.CreateTaskResultValue(value);
-
-            var resultValueAfter = await resultHasError.ToResultBindValueBindAsync(resultValue);
-
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой в результирующем ответе
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueBindAsync_HasErrorsBind()
-        {
-            var resultNoError = ResultErrorFactory.CreateTaskResultError();
-            var error = ErrorData.CreateErrorTest();
-            var resultValue = ResultValueFactory.CreateTaskResultValueError<string>(error);
-
-            var resultValueAfter = await resultNoError.ToResultBindValueBindAsync(resultValue);
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultValueBindBindAsync_HasErrorsBindInitial()
-        {
-            var error = ErrorData.CreateErrorTest();
-            var resultHasError = ResultErrorFactory.CreateTaskResultError(error);
-            var errors = ErrorData.CreateErrorListTwoTest();
-            var resultValue = ResultValueFactory.CreateTaskResultValueError<string>(errors);
-
-            var resultValueAfter = await resultHasError.ToResultBindValueBindAsync(resultValue);
-
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-    }
-}
