+++ ResultFunctionalXUnit/FunctionalExtensions/Async/ResultExtension/ResultErrorTest/ToResultErrorTaskAsyncExtensionsTest.cs
@@ -17,75 +17,6 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async.ResultExtension.Resul
     public class ToResultErrorTaskAsyncExtensionsTest
     {
         /// <summary>
-        /// Вернуть результирующий ответ со значением без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueTaskAsync_OkStatus()
-        {
-            var resultNoError = ResultErrorFactory.CreateTaskResultError();
-            const string value = "OkStatus";
-            var resultValue = new ResultValue<string>(value);
-
-            var resultValueAfter = await resultNoError.ToResultBindValueTaskAsync(resultValue);
-
-            Assert.True(resultValueAfter.OkStatus);
-            Assert.Equal(value, resultValueAfter.Value);
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueTaskAsync_HasErrors()
-        {
-            var error = ErrorData.CreateErrorTest();
-            var resultHasError = ResultErrorFactory.CreateTaskResultError(error);
-            const string value = "BadStatus";
-            var resultValue = new ResultValue<string>(value);
-
-            var resultValueAfter = await resultHasError.ToResultBindValueTaskAsync(resultValue);
-
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой в результирующем ответе
-        /// </summary>      
-        [Fact]
-        public async Task ToResultBindValueTaskAsync_HasErrorsBind()
-        {
-            var resultNoError = ResultErrorFactory.CreateTaskResultError();
-            var error = ErrorData.CreateErrorTest();
-            var resultValue = new ResultValue<string>(error);
-
-            var resultValueAfter = await resultNoError.ToResultBindValueTaskAsync(resultValue);
-
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultValueBindTaskAsync_HasErrorsBindInitial()
-        {
-            var error = ErrorData.CreateErrorTest();
-            var resultHasError = ResultErrorFactory.CreateTaskResultError(error);
-            var errors = ErrorData.CreateErrorListTwoTest();
-            var resultValue = new ResultValue<string>(errors);
-
-            var resultValueAfter = await resultHasError.ToResultBindValueTaskAsync(resultValue);
-
-            Assert.True(resultValueAfter.HasErrors);
-            Assert.Single(resultValueAfter.Errors);
-            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
-        }
-
-        /// <summary>
         /// Преобразовать в результирующий ответ
         /// </summary>
         [Fact]
