+++ ResultFunctionalXUnit/FunctionalExtensions/Sync/ResultExtension/ResultValueTest/ToResultValueExtensionsTest.cs
@@ -177,5 +177,107 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Sync.ResultExtension.Result
             Assert.True(result.HasErrors);
             Assert.True(result.Errors.First().Equals(initialError));
         }
+
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением без ошибок
+        /// </summary>      
+        [Fact]
+        public void ToResultValueBind_OkStatus()
+        {
+            var resultNoError = new ResultError();
+            const string value = "OkStatus";
+            var resultValue = new ResultValue<string>(value);
+
+            var resultValueAfter = resultNoError.ToResultBindValue(resultValue);
+
+            Assert.True(resultValueAfter.OkStatus);
+            Assert.Equal(value, resultValueAfter.Value);
+        }
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением с ошибкой
+        /// </summary>      
+        [Fact]
+        public void ToResultValueBind_HasErrors()
+        {
+            var error = CreateErrorTest();
+            var resultHasError = new ResultError(error);
+            const string value = "BadStatus";
+            var resultValue = new ResultValue<string>(value);
+
+            var resultValueAfter = resultHasError.ToResultBindValue(resultValue);
+
+            Assert.True(resultValueAfter.HasErrors);
+            Assert.Single(resultValueAfter.Errors);
+            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
+        }
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением без ошибок
+        /// </summary>      
+        [Fact]
+        public void ToResultValueFromResultError_OkStatus()
+        {
+            var resultNoError = new ResultError();
+            const string value = "OkStatus";
+
+            var resultValue = resultNoError.ToResultValue(value);
+
+            Assert.True(resultValue.OkStatus);
+            Assert.Equal(value, resultValue.Value);
+        }
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением с ошибкой
+        /// </summary>      
+        [Fact]
+        public void ToResultValueFromResultError_HasErrors()
+        {
+            var error = CreateErrorTest();
+            var resultHasError = new ResultError(error);
+            const string value = "BadStatus";
+
+            var resultValue = resultHasError.ToResultValue(value);
+
+            Assert.True(resultValue.HasErrors);
+            Assert.Single(resultValue.Errors);
+            Assert.True(error.Equals(resultValue.Errors.Last()));
+        }
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением с ошибкой в результирующем ответе
+        /// </summary>      
+        [Fact]
+        public void ToResultValueBind_HasErrorsBind()
+        {
+            var resultNoError = new ResultError();
+            var error = CreateErrorTest();
+            var resultValue = new ResultValue<string>(error);
+
+            var resultValueAfter = resultNoError.ToResultBindValue(resultValue);
+
+            Assert.True(resultValueAfter.HasErrors);
+            Assert.Single(resultValueAfter.Errors);
+            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
+        }
+
+        /// <summary>
+        /// Вернуть результирующий ответ со значением с ошибкой
+        /// </summary>      
+        [Fact]
+        public void ToResultValueBind_HasErrorsBindInitial()
+        {
+            var error = CreateErrorTest();
+            var resultHasError = new ResultError(error);
+            var errors = CreateErrorListTwoTest();
+            var resultValue = new ResultValue<string>(errors);
+
+            var resultValueAfter = resultHasError.ToResultBindValue(resultValue);
+
+            Assert.True(resultValueAfter.HasErrors);
+            Assert.Single(resultValueAfter.Errors);
+            Assert.True(error.Equals(resultValueAfter.Errors.Last()));
+        }
     }
 }
